Add missing effects to modifier array.
Change AMP_POSITION to AMP_A_POSITION.

Add AMP_B_POSITION to the modifier array.

Add AMP_PRESENCE to the modifier array and move it to follow Treble
in the Equalizer effect.

Add EQ_B_POSITION settings to the modifier array.

Change the formatting of the debug messages to print more
readable columns.