Enforce hysteresis on modifier list requests.
With the tone and effect libraries implemented on rp355, it appears that
the device will send multiple NOTIFY_MODIFIER_GROUP_CHANGED messages
in response to selecting a new tone or effect library.

If we receive a NOTIFY_MODIFIER_GROUP_CHANGED, but haven't yet processed
the RECEIVE_MODIFIER_LINKABLE_LIST from the previous notification,
we will drop the most recent.

This greatly reduces the device churn and time it takes to apply a new library
setting.